pyproject.toml
[project]
name = "cryptographic-ciphers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.4.2",
    "pycryptodome>=3.23.0",
]
